sb動作確認
