+++ src/stories/components/atoms/Button.stories.tsx
@@ -1,0 +1,35 @@
+import type { Meta, StoryObj } from "@storybook/react";
+
+import { Button } from "./Button";
+
+const meta = {
+  title: "Button",
+  component: Button,
+  parameters: {
+    layout: "centered",
+  },
+  tags: ["autodocs"],
+  argTypes: {
+    backgroundColor: { control: "color" },
+  },
+} satisfies Meta<typeof Button>;
+
+export default meta;
+
+type Story = StoryObj<typeof meta>;
+
+export const Primary: Story = {
+  args: {
+    primary: true,
+    label: "PrimaryButton",
+    color: "red",
+  },
+};
+
+export const Secondary: Story = {
+  args: {
+    primary: false,
+    label: "SecondaryButton",
+    color: "red",
+  },
+};
